+++ src/utils/appDirs.ts
@@ -6,21 +6,10 @@
 
 import path from 'path';
 
-import { OFFICIAL } from '../../ipc/sources';
+import launcherConfig from './launcherConfig';
 import { packageJsonApp } from './packageJson';
 
-const getUserDataDir = () => {
-    const argv = process.argv;
-    const userDataDir = argv.findIndex(arg =>
-        arg.startsWith('--user-data-dir')
-    );
-
-    if (userDataDir === -1) {
-        throw new Error('Could not find --user-data-dir argument');
-    }
-
-    return argv[userDataDir].split('=')[1];
-};
+const getUserDataDir = () => launcherConfig().userDataDir;
 
 /**
  * Get the filesystem path of the currently loaded app.
@@ -56,19 +45,4 @@ const getAppDataDir = () =>
  */
 const getAppLogDir = () => path.join(getAppDataDir(), 'logs');
 
-const getAppSource = () => {
-    const source = path
-        .parse(path.resolve(getAppDir(), '../'))
-        .base.toLowerCase();
-
-    return source === 'node_modules' ? OFFICIAL : source;
-};
-
-export {
-    getAppDir,
-    getAppFile,
-    getAppDataDir,
-    getAppLogDir,
-    getUserDataDir,
-    getAppSource,
-};
+export { getAppDir, getAppFile, getAppDataDir, getAppLogDir, getUserDataDir };